完善文章模块的功能，新增照片墙功能 (#6)
* chore: 添加配置和修改功能

1.将获取统计数据改为根据传进来的天数获取
2.添加自定义业务异常

* fix: 修复获取统计数据时，没有当月数据的bug

* chore: 配置文件上传

1.记录用户id
2.修改文件访问路径

* chore: 新增和修改日记的功能

1.日记可以进行审核
2.新增获取所有状态的日记接口
3.在非管理页面时只能获取审核通过的日记

* chore: 完善日记管理功能以及一些配置

1.日记需要进行审核才能展示（审核后远程调用用户服务发送邮件通知
2.提取邮箱枚举类到公共代码中
3.新增拒绝和通过日记的邮箱策略

* feat: 新增添加文章分组接口以及统计碎片数量

* feat: 新增文章的一些功能

1.新增根据用户最近10条点赞信息获取推荐文章
2.新增获取分组时，同时返回分组所拥有的文章数

* feat: 新增文章搜索接口

1.能根据时间、阅读量、点赞进行排序
2.能搜索描述和标题
2.能根据标签进行检索

* fix: 修复文章添加异常

1.文章分组类型没修改

* fix: 修复推荐文章和获取日记接口的bug

1.修复推荐文章除0异常
2.添加日记时，没有初始化审核状态

* chore: 完善图片上传功能

1.根据权限限制用户所能上传的总大小
2.图片上传额外生成缩略图

* feat: 新增图片模块的一些功能

1.新增返回每个用户图片数量的接口
2.新增根据用户获取图片列表的接口

* feat: 新增图片模块的一些功能

1.新增返回每个用户图片数量的接口
2.新增根据用户获取图片列表的接口

* fix: 修复将用户名作为昵称返回的bug

* feat: 新增照片模块的一些接口

1.新增图片设为公开和私密接口
2.新增图片删除接口
3.新增获取自己图片的接口

+++ text-service/src/main/java/com/llpy/textservice/entity/dto/ArticleDto.java
@@ -41,5 +41,5 @@ public class ArticleDto {
     /**
      * 组id
      */
-    private List<Integer> groupIds;
+    private List<Long> groupIds;
 }
